Quick bug fix

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -2,16 +2,16 @@
 
 
 /**
+ * listint_len - Function that returns number of elements in list
+ * @h: Our linked list
  *
- *
- *
- *
+ * Return: number of elements in list
  */
 
 size_t listint_len(const listint_t *h)
 {
 
-int i;
+size_t i;
 
 for (i = 0; h; i++)
 {
